+++ sdk/eventgrid/Microsoft.Azure.WebJobs.Extensions.EventGrid/src/TriggerBinding/BindingType.cs
@@ -1,0 +1,12 @@
+﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+namespace Microsoft.Azure.WebJobs.Extensions.EventGrid
+{
+    internal enum BindingType
+    {
+        Unknown,
+        EventGridEvent,
+        CloudEvent
+    }
+}