+++ sdk/resourcemanager/Azure.ResourceManager/src/ResourceListOperations.cs
@@ -29,13 +29,7 @@ namespace Azure.ResourceManager.Core
             int? top = null,
             CancellationToken cancellationToken = default)
         {
-            return ListAtContextInternal(
-                resourceGroup,
-                resourceGroup.Id.Name,
-                resourceFilters,
-                expand,
-                top,
-                cancellationToken);
+            return resourceGroup.GetGenericResources(resourceFilters?.ToString(), expand, top, cancellationToken);
         }
 
         /// <summary>
@@ -55,13 +49,7 @@ namespace Azure.ResourceManager.Core
             int? top = null,
             CancellationToken cancellationToken = default)
         {
-            return ListAtContextInternalAsync(
-                resourceGroup,
-                resourceGroup.Id.Name,
-                resourceFilters,
-                expand,
-                top,
-                cancellationToken);
+            return resourceGroup.GetGenericResourcesAsync(resourceFilters?.ToString(), expand, top, cancellationToken);
         }
 
         /// <summary>
@@ -81,13 +69,7 @@ namespace Azure.ResourceManager.Core
             int? top = null,
             CancellationToken cancellationToken = default)
         {
-            return ListAtContextInternal(
-                subscription,
-                null,
-                resourceFilters,
-                expand,
-                top,
-                cancellationToken);
+            return subscription.GetGenericResources(resourceFilters?.ToString(), expand, top, cancellationToken);
         }
 
         /// <summary>
@@ -107,72 +89,7 @@ namespace Azure.ResourceManager.Core
             int? top = null,
             CancellationToken cancellationToken = default)
         {
-            return ListAtContextInternalAsync(
-                subscription,
-                null,
-                resourceFilters,
-                expand,
-                top,
-                cancellationToken);
-        }
-
-        private static GenericResourceCollection GetGenericResourceCollection(ArmResource resourceOperations)
-        {
-            return new GenericResourceCollection(new ClientContext(resourceOperations.ClientOptions, resourceOperations.Credential, resourceOperations.BaseUri, resourceOperations.Pipeline), resourceOperations.Id);
-        }
-
-        private static AsyncPageable<GenericResource> ListAtContextInternalAsync(
-            ArmResource resourceOperations,
-            string scopeFilter,
-            ResourceFilterCollection resourceFilters = null,
-            string expand = null,
-            int? top = null,
-            CancellationToken cancellationToken = default)
-        {
-            var restClient = GetGenericResourceCollection(resourceOperations);
-            AsyncPageable<GenericResource> result;
-            if (scopeFilter == null)
-            {
-                result = restClient.GetAllAsync(resourceFilters?.ToString(), expand, top, cancellationToken);
-            }
-            else
-            {
-                result = restClient.GetByResourceGroupAsync(
-                    scopeFilter,
-                    resourceFilters?.ToString(),
-                    expand,
-                    top,
-                    cancellationToken);
-            }
-
-            return result;
-        }
-
-        private static Pageable<GenericResource> ListAtContextInternal(
-            ArmResource resourceOperations,
-            string scopeFilter = null,
-            ResourceFilterCollection resourceFilters = null,
-            string expand = null,
-            int? top = null,
-            CancellationToken cancellationToken = default)
-        {
-            var restClient = GetGenericResourceCollection(resourceOperations);
-            Pageable<GenericResource> result;
-            if (scopeFilter == null)
-            {
-                result = restClient.GetAll(resourceFilters?.ToString(), expand, top, cancellationToken);
-            }
-            else
-            {
-                result = restClient.GetByResourceGroup(
-                    scopeFilter,
-                    resourceFilters?.ToString(),
-                    expand,
-                    top,
-                    cancellationToken);
-            }
-
-            return result;
+            return subscription.GetGenericResourcesAsync(resourceFilters?.ToString(), expand, top, cancellationToken);
         }
     }
 }