+++ sdk/resourcemanager/Azure.ResourceManager/tests/Scenario/ResourceGroupCollectionTests.cs
@@ -64,9 +64,8 @@ namespace Azure.ResourceManager.Tests
             Assert.AreEqual(rgName, rg.Data.Name);
 
             Assert.Throws<ArgumentNullException>(() => { subscription.GetResourceGroups().Construct(null); });
-            Assert.ThrowsAsync<ArgumentException>(async () => _ = await subscription.GetResourceGroups().CreateOrUpdateAsync(null, new ResourceGroupData(AzureLocation.WestUS2)));
-            Assert.ThrowsAsync<ArgumentException>(async () => _ = await subscription.GetResourceGroups().CreateOrUpdateAsync(" ", new ResourceGroupData(AzureLocation.WestUS2)));
-            Assert.ThrowsAsync<ArgumentNullException>(async () => _ = await subscription.GetResourceGroups().CreateOrUpdateAsync(rgName, null));
+            Assert.ThrowsAsync<ArgumentNullException>(async () => _ = await subscription.GetResourceGroups().CreateOrUpdateAsync(true, null, new ResourceGroupData(AzureLocation.WestUS2)));
+            Assert.ThrowsAsync<ArgumentNullException>(async () => _ = await subscription.GetResourceGroups().CreateOrUpdateAsync(true, rgName, null));
         }
 
         [TestCase]
@@ -79,19 +78,14 @@ namespace Azure.ResourceManager.Tests
             ResourceGroup rg = await rgOp.WaitForCompletionAsync();
             Assert.AreEqual(rgName, rg.Data.Name);
 
-            Assert.ThrowsAsync<ArgumentException>(async () =>
+            Assert.ThrowsAsync<ArgumentNullException>(async () =>
             {
-                var rgOp = await subscription.GetResourceGroups().CreateOrUpdateAsync(null, new ResourceGroupData(AzureLocation.WestUS2), false);
-                _ = await rgOp.WaitForCompletionAsync();
-            });
-            Assert.ThrowsAsync<ArgumentException>(async () =>
-            {
-                var rgOp = await subscription.GetResourceGroups().CreateOrUpdateAsync(" ", new ResourceGroupData(AzureLocation.WestUS2), false);
+                var rgOp = await subscription.GetResourceGroups().CreateOrUpdateAsync(false, null, new ResourceGroupData(AzureLocation.WestUS2));
                 _ = await rgOp.WaitForCompletionAsync();
             });
             Assert.ThrowsAsync<ArgumentNullException>(async () =>
             {
-                var rgOp = await subscription.GetResourceGroups().CreateOrUpdateAsync(rgName, null, false);
+                var rgOp = await subscription.GetResourceGroups().CreateOrUpdateAsync(false, rgName, null);
                 _ = await rgOp.WaitForCompletionAsync();
             });
         }
@@ -120,16 +114,16 @@ namespace Azure.ResourceManager.Tests
 
         [TestCase]
         [RecordedTest]
-        public async Task CheckIfExists()
+        public async Task Exists()
         {
             Subscription subscription = await Client.GetDefaultSubscriptionAsync().ConfigureAwait(false);
             var rgName = Recording.GenerateAssetName("testRg-");
             var rgOp = await subscription.GetResourceGroups().Construct(AzureLocation.WestUS2).CreateOrUpdateAsync(rgName);
             ResourceGroup rg = rgOp.Value;
-            Assert.IsTrue(await subscription.GetResourceGroups().CheckIfExistsAsync(rgName));
-            Assert.IsFalse(await subscription.GetResourceGroups().CheckIfExistsAsync(rgName + "1"));
+            Assert.IsTrue(await subscription.GetResourceGroups().ExistsAsync(rgName));
+            Assert.IsFalse(await subscription.GetResourceGroups().ExistsAsync(rgName + "1"));
 
-            Assert.ThrowsAsync<ArgumentNullException>(async () => _ = await subscription.GetResourceGroups().CheckIfExistsAsync(null));
+            Assert.ThrowsAsync<ArgumentNullException>(async () => _ = await subscription.GetResourceGroups().ExistsAsync(null));
         }
 
         [TestCase]
@@ -146,14 +140,14 @@ namespace Azure.ResourceManager.Tests
             var response = await subscription.GetResourceGroups().GetIfExistsAsync(rgName + "1");
             Assert.IsNull(response.Value);
 
-            Assert.ThrowsAsync<ArgumentNullException>(async () => _ = await subscription.GetResourceGroups().CheckIfExistsAsync(null));
+            Assert.ThrowsAsync<ArgumentNullException>(async () => _ = await subscription.GetResourceGroups().ExistsAsync(null));
         }
 
         [RecordedTest]
         public async Task EnumerableInterface()
         {
             Subscription subscription = await Client.GetDefaultSubscriptionAsync();
-            var rgOp = await subscription.GetResourceGroups().CreateOrUpdateAsync(Recording.GenerateAssetName("testRg-"), new ResourceGroupData(AzureLocation.WestUS2));
+            var rgOp = await subscription.GetResourceGroups().CreateOrUpdateAsync(true, Recording.GenerateAssetName("testRg-"), new ResourceGroupData(AzureLocation.WestUS2));
             ResourceGroup rg = rgOp.Value;
             int count = 0;
             await foreach(var rgFromList in subscription.GetResourceGroups())