+++ sdk/monitor/Azure.Monitor.Query/CHANGELOG.md
@@ -9,6 +9,7 @@ Added the following methods to `MonitorQueryModelFactory`:
 - `LogsTableRow`
 - `MetricsQueryResult`
 - `MetricValue`
+- `MetricTimeSeriesElement`
 
 ## 1.0.1 (2021-11-09)